// src/lib/api/types.ts
export interface ImageRef {
  url: string
  width: number
  height: number
  quality?: string
}
// ...
export interface VideoSummary {
  type: 'video' | 'shortVideo'
  title: string
  videoId: string
  author: string
  authorId: string
  authorUrl?: string
  videoThumbnails: ImageRef[]
  description?: string
  descriptionHtml?: string
  viewCount: number
  viewCountText?: string
  published?: number
  publishedText?: string
  lengthSeconds: number
  liveNow?: boolean
  premium?: boolean
  isUpcoming?: boolean
  premiereTimestamp?: number
}

export interface ChannelSummary {
  type: 'channel'
  author: string
  authorId: string
  authorUrl: string
  authorThumbnails: ImageRef[]
  autoGenerated?: boolean
  subCount: number
  videoCount: number
  description?: string
  descriptionHtml?: string
}

export interface ChannelDetails extends ChannelSummary {
  authorBanners: ImageRef[]
  totalViews?: number
  joined?: number
  isFamilyFriendly?: boolean
  allowedRegions?: string[]
  tabs?: string[]
  latestVideos?: VideoSummary[]
  relatedChannels?: ChannelSummary[]
}

export interface PlaylistSummary {
  type: 'playlist'
  title: string
  playlistId: string
  author: string
  authorId: string
  authorUrl?: string
  playlistThumbnail?: string
  videoCount: number
  videos?: VideoSummary[]
}

export interface PlaylistDetails extends PlaylistSummary {
  description?: string
  descriptionHtml?: string
  viewCount?: number
  updated?: number
  isListed?: boolean
  videos: VideoSummary[]
}